move chkpnt booter component into llbooter
update documentation of booter
