--- src/components/implementation/tests/chkpnt/Makefile
@@ -1,18 +1,0 @@
-# Required variables used to drive the compilation process. It is OK
-# for many of these to be empty.
-#
-# The set of interfaces that this component exports for use by other
-# components. This is a list of the interface names.
-INTERFACE_EXPORTS = init addr
-# The interfaces this component is dependent on for compilation (this
-# is a list of directory names in interface/)
-INTERFACE_DEPENDENCIES = init
-# The library dependencies this component is reliant on for
-# compilation/linking (this is a list of directory names in lib/)
-LIBRARY_DEPENDENCIES = component crt
-# Note: Both the interface and library dependencies should be
-# *minimal*. That is to say that removing a dependency should cause
-# the build to fail. The build system does not validate this
-# minimality; that's on you!
-
-include Makefile.subsubdir
