--- src/components/implementation/tests/chkpnt/chkpnt.h
@@ -1,18 +1,0 @@
-// #ifndef CHKPNT
-// #define CHKPNT
-
-// #define MAX_NUM_CHKPNTS 4
-
-// struct crt_chkpt {
-//     struct crt_comp *c;
-//     /* some kind of pointer to the memory */
-
-// };
-
-// int crt_chkpt_create(struct crt_chkpt *cp, struct crt_comp *c);
-// int crt_chkpt_restore(struct crt_chkpt *cp, struct crt_comp *c);
-
-// #endif //CHKPNT
-
-
-
