--- src/components/implementation/tests/chkpnt/doc.md
@@ -1,7 +1,0 @@
-## chkpnt
-
-### Description
-
-This is the first iteration of a component that will be invoked to checkpoint another component. Right now, just trying to create multiple of one component
-
-### Usage and Assumptions
